# Use Bini's algorithm to select initial values in acb_poly_find_roots