--- a/src/acb_poly/find_roots.c
+++ b/src/acb_poly/find_roots.c
@@ -1,5 +1,5 @@
 /*
-    Copyright (C) 2012 Fredrik Johansson
+    Copyright (C) 2012, 2024 Fredrik Johansson
 
     This file is part of FLINT.
 
@@ -9,7 +9,9 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
+#include <math.h>
 #include "ulong_extras.h"
+#include "arf.h"
 #include "fmpq.h"
 #include "acb_poly.h"
 
@@ -44,29 +46,102 @@ _acb_get_rad_mag(const acb_t z)
     return FLINT_MAX(rm, im);
 }
 
-void
-_acb_poly_roots_initial_values(acb_ptr roots, slong deg, slong prec)
+/* Compute res[0, ..., n-1] = {i} indexing the convex hull of {i, y[i]}. */
+static slong convex_hull(slong * res, const double * y, slong len)
 {
-    slong i;
+    slong i, n = 0;
 
-    fmpq_t q;
-    fmpq_init(q);
-
-    fmpq_set_si(q, 4, 10);
-    arb_set_fmpq(acb_realref(roots + 0), q, prec);
-    mag_zero(arb_radref(acb_realref(roots + 0)));
-
-    fmpq_set_si(q, 9, 10);
-    arb_set_fmpq(acb_imagref(roots + 0), q, prec);
-    mag_zero(arb_radref(acb_imagref(roots + 0)));
-    fmpq_clear(q);
-
-    for (i = 1; i < deg; i++)
+    for (i = 0; i < len; i++)
     {
-        acb_mul(roots + i, roots + i - 1, roots + 0, prec);
-        mag_zero(arb_radref(acb_realref(roots + i)));
-        mag_zero(arb_radref(acb_imagref(roots + i)));
+        while (n >= 2 && (res[n - 2] - res[n - 1]) * (y[i] - y[res[n - 1]])
+                <= (i - res[n - 1]) * (y[res[n - 2]] - y[res[n - 1]]))
+            n--;
+
+        res[n] = i;
+        n++;
     }
+
+    return n;
+}
+
+void
+_acb_poly_roots_initial_values(acb_ptr roots, acb_srcptr poly, slong deg, slong prec)
+{
+    double * alog;
+    mag_ptr amag;
+    mag_t r;
+    arf_t ar;
+    slong i, j;
+    slong * ki, num, m, total;
+    double theta;
+    acb_t cmid;
+    acb_ptr ri;
+
+    amag = _mag_vec_init(deg + 1);
+    alog = flint_malloc(sizeof(double) * (deg + 1));
+    ki = flint_malloc((deg + 1) * sizeof(slong));
+    mag_init(r);
+    arf_init(ar);
+    acb_init(cmid);  /* shallow only; will not be cleared */
+
+    for (i = 0; i <= deg; i++)
+    {
+        /* shallow midpoint */
+        *arb_midref(acb_realref(cmid)) = *arb_midref(acb_realref(poly + i));
+        *arb_midref(acb_imagref(cmid)) = *arb_midref(acb_imagref(poly + i));
+        acb_get_mag(amag + i, cmid);
+
+        /* todo: mag_get_d_log2_approx is not very precise; probably
+           this does not matter though? */
+        alog[i] = mag_get_d_log2_approx(amag + i);
+    }
+
+    num = convex_hull(ki, alog, deg + 1);
+    total = 0;
+
+    for (i = 1; i < num; i++)
+    {
+        /* multiplicity */
+        m = ki[i] - ki[i - 1];
+
+        /* radius */
+        mag_div(r, amag + ki[i - 1], amag + ki[i]);
+        mag_root(r, r, m);
+
+        if (mag_is_zero(r))
+            mag_set_ui_2exp_si(r, 1, -prec);
+        if (mag_is_inf(r))
+            mag_set_ui_2exp_si(r, 1, prec);
+
+        arf_set_mag(ar, r);
+
+#if 0
+        flint_printf("initial values: radius %{mag} multiplicity %wd\n", r, m);
+#endif
+
+        /* pick m points on the circle with radius r */
+        for (j = 0; j < m; j++)
+        {
+            theta = 6.283185307179586 * ((j + 1.0) / m + (double) i / deg) + 0.577216;
+
+            ri = roots + total;
+            acb_zero(ri);
+            arf_set_d(arb_midref(acb_realref(ri)), cos(theta));
+            arf_set_d(arb_midref(acb_imagref(ri)), sin(theta));
+            arf_mul(arb_midref(acb_realref(ri)), arb_midref(acb_realref(ri)), ar, MAG_BITS, ARF_RND_DOWN);
+            arf_mul(arb_midref(acb_imagref(ri)), arb_midref(acb_imagref(ri)), ar, MAG_BITS, ARF_RND_DOWN);
+            total++;
+        }
+    }
+
+    if (total != deg)
+        flint_abort();
+
+    _mag_vec_clear(amag, deg + 1);
+    flint_free(alog);
+    flint_free(ki);
+    mag_clear(r);
+    arf_clear(ar);
 }
 
 slong
@@ -102,7 +177,7 @@ _acb_poly_find_roots(acb_ptr roots,
     }
 
     if (initial == NULL)
-        _acb_poly_roots_initial_values(roots, deg, prec);
+        _acb_poly_roots_initial_values(roots, poly, deg, prec);
     else
         _acb_vec_set(roots, initial, deg);